IconLogo.dart file added and other modifications done
